# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hickory Docs API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/billing/checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "operationId": "checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stripe Checkout session url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/webhook": {
      "post": {
        "tags": [
          "billing"
        ],
        "operationId": "webhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "in": "header",
            "description": "Stripe webhook signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "event acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutOut": {
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "properties": {
          "checkout_url": {
            "type": "string"
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": [
          "price_key"
        ],
        "properties": {
          "price_key": {
            "type": "string"
          }
        }
      },
...
      "WebhookOut": {
        "type": "object",
        "required": [
          "received"
        ],
        "properties": {
          "duplicate": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "received": {
            "type": "boolean"
          }
        }
      }
    }
  }
}